+++ java/test/org/openqa/selenium/grid/node/local/LocalNodeTest.java
@@ -429,13 +429,6 @@ class LocalNodeTest {
   }
 
   @Test
-  void extractsFileNameFromRequestUri() {
-    assertThat(node.extractFileName("/session/1234/se/files/logo.png")).isEqualTo("logo.png");
-    assertThat(node.extractFileName("/session/1234/se/files/файл+with+tähtedega.png"))
-        .isEqualTo("файл+with+tähtedega.png");
-  }
-
-  @Test
   void commandInterceptorIsCalledForEachWebDriverCommand() throws URISyntaxException {
     Tracer tracer = DefaultTestTracer.createTracer();
     EventBus bus = new GuavaEventBus();